refactor: artist 폴더 하위에 domain 폴더를 가지도록 변경
