+++ src/main/java/com/ziine/admin/auth/application/JwtService.java
@@ -9,9 +9,9 @@ import java.util.Optional;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
-@Component
+@Service
 @EnableConfigurationProperties(JwtProperties.class)
 public class JwtService {
