+++ src/main/java/com/ziine/domains/artwork/application/ArtworkFacade.java
@@ -1,8 +1,8 @@
 package com.ziine.domains.artwork.application;
 
 import com.ziine.domains.artist.application.ArtistService;
+import com.ziine.domains.artist.domain.entity.ArtistEntity;
 import com.ziine.domains.artist.dto.ArtistPersistDto;
-import com.ziine.domains.artist.entity.ArtistEntity;
 import com.ziine.domains.artwork.domain.entity.ArtworkEntity;
 import com.ziine.domains.artwork.dto.ArtworkPersistDto;
 import com.ziine.domains.artwork.dto.request.ArtworkPersistRequestDto;
